+++ src/Generator/Model/FederationRepresentationResolverGeneratorInterface.php
@@ -1,0 +1,6 @@
+<?php
+
+namespace Axtiva\FlexibleGraphql\Generator\Model;
+
+interface FederationRepresentationResolverGeneratorInterface extends ModelGeneratorInterface
+{}